DDF-5131 moved session lifetime setting to the session factory
