+++ distribution/ddf-common/src/main/resources-filtered/etc/application-definitions/security-services.json
@@ -6,6 +6,7 @@
     "mvn:ddf.platform.security/security-rest-authn-impl/${project.version}",
     "mvn:ddf.platform.security/security-rest-authn-endpoint/${project.version}",
     "mvn:ddf.security.core/security-core-impl/${project.version}",
+    "mvn:ddf.security.core/security-core-services/${project.version}",
     "mvn:ddf.security.expansion/security-expansion-metacardattr-map/${project.version}",
     "mvn:ddf.security.expansion/security-expansion-userattr-map/${project.version}",
     "mvn:ddf.security.filter/security-filter-login/${project.version}",
