--- platform/security/filter/security-filter-login/src/main/resources/OSGI-INF/metatype/metatype.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-/**
- * Copyright (c) Codice Foundation
- *
- * This is free software: you can redistribute it and/or modify it under the terms of the GNU Lesser General Public License as published by the Free Software Foundation, either
- * version 3 of the License, or any later version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
- * See the GNU Lesser General Public License for more details. A copy of the GNU Lesser General Public License is distributed along with this program and can be found at
- * <http://www.gnu.org/licenses/lgpl.html>.
- *
- **/
-
- -->
-<metatype:MetaData xmlns:metatype="http://www.osgi.org/xmlns/metatype/v1.0.0">
-
-    <OCD name="Session" id="org.codice.ddf.security.filter.login.Session">
-        <AD description="Specifies the length of inactivity (in minutes) between client requests before the servlet container will invalidate the session (this applies to all client sessions).
-            This value must be 2 minutes or greater, as users are warned when only 1 minute remains. If a value of less than 2 minutes is used, the timeout is set to the default time of 31 minutes."
-            name="Session Timeout (in minutes)" id="expirationTime" required="true" type="Integer"
-            default="31"/>
-    </OCD>
-
-    <Designate pid="org.codice.ddf.security.filter.login.Session">
-        <Object ocdref="org.codice.ddf.security.filter.login.Session"/>
-    </Designate>
-
-</metatype:MetaData>
